pyproject.toml
[project]
name = "leadgen-mcp"
version = "0.1.0"
description = "Serveur MCP pour pipeline lead gen PME QC — wraps Supabase, Google Places, Apollo, Instantly, Vapi"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.6.0",
]

[project.scripts]
leadgen-mcp = "server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
